+++ src/components/home/WeeklyDate/WeeklyDate.tsx
@@ -1,29 +1,21 @@
 import { useState, useEffect } from 'react';
 import DateItem from './DateItem/DateItem';
 import getWeekDates, { WeekDates } from '@/utils/getWeekDates';
-import getFormattedDate from '@/utils/getFormattedDate';
 import { Carousel, CarouselContent, CarouselItem } from '@/components/common/ui/carousel';
 import getWeekText from '@/utils/getWeekText';
 import useHomePageStore from '@/store/useHomePageStore';
 import { type CarouselApi } from '@/components/common/ui/carousel';
-import { getHouseworks } from '@/services/housework/getHouseworks';
-import { PAGE_SIZE } from '@/constants/common';
-import { useParams } from 'react-router-dom';
 
 const WeeklyDate = () => {
   const [activeWeek, setActiveWeek] = useState(new Date());
-  const [activeDate, setActiveDate] = useState<string>(getFormattedDate(new Date()));
   const [currWeek, setCurrWeek] = useState<WeekDates[]>([]);
   const [api, setApi] = useState<CarouselApi>();
   const [current, setCurrent] = useState(0);
-  const [pageNumber, setPageNumber] = useState(0);
-  const { setWeekText, setHouseworks } = useHomePageStore();
-  const { channelId } = useParams();
+  const { setWeekText, activeDate, setActiveDate } = useHomePageStore();
 
   useEffect(() => {
     const weekDates = getWeekDates(activeWeek);
     setCurrWeek(weekDates);
-    fetchHouseworks(activeDate);
   }, []);
 
   useEffect(() => {
@@ -42,22 +34,8 @@ const WeeklyDate = () => {
     });
   }, [api, current]);
 
-  const fetchHouseworks = async (date: string) => {
-    const newChannelId = Number(channelId);
-    const getHouseworksResult = await getHouseworks({
-      channelId: newChannelId,
-      targetDate: date,
-      pageNumber: pageNumber,
-      pageSize: PAGE_SIZE,
-    });
-    console.log(getHouseworksResult);
-    setHouseworks(getHouseworksResult.result.responses);
-  };
-
   const handleActiveDate = async (date: string) => {
     setActiveDate(date);
-    // 추후에 해당 날짜 집안일 리스트 불러오기
-    fetchHouseworks(date);
   };
 
   const changeWeek = (direction: 'next' | 'previous') => {
